Fix CPS1/2 mapping
This fix to mapping allows XInput arcade sticks and gamepads to map directly to modern Street Fighter controls.

Non-fighting games on CPS1/2 will probably still require custom mappings.

Rationale: Fighting games are the bulk of the CPS1/2 library, and it ends up being easier to do per-game for the rest instead of the other way around.